Re-arranged HUD

+++ src/main/java/calc/train/service/CalculateSimple.java
@@ -6,16 +6,16 @@ import org.springframework.stereotype.Service;
 @Service
 public class CalculateSimple {
 
-	public int add(Calculator model){
-        return (int) (model.getOperandL() + model.getOperandR());
+	public double add(Calculator model){
+        return (model.getOperandL() + model.getOperandR());
     }
 
-    public int subtract(Calculator model){
-        return (int) (model.getOperandL() - model.getOperandR());
+    public double subtract(Calculator model){
+        return (model.getOperandL() - model.getOperandR());
     }
 
-    public int multiply(Calculator model){
-        return (int) (model.getOperandL() * model.getOperandR());
+    public double multiply(Calculator model){
+        return (model.getOperandL() * model.getOperandR());
     }
 
     public double divide(Calculator model){
